Minor formatting fix

+++ src/rawposix/src/fs_calls.rs
@@ -1603,9 +1603,7 @@ pub fn getdents_syscall(
         return syscall_error(Errno::EFAULT, "getdents", "Invalid Cage ID");
     }
 
-    let ret = unsafe {
-        libc::syscall(libc::SYS_getdents64 as libc::c_long, kernel_fd, dirp, count) as i64
-    }
+    let ret = unsafe { libc::syscall(libc::SYS_getdents64 as libc::c_long, kernel_fd, dirp, count) }
 
     ret 
 }
